Various vmservice/observatory fixes:
- get rid of double lines on vm view page

- show eval errors in monospace font

- make eval highlight color consistent

- fix broken navbar on script view page

- show list of all libs for an isolate

- use two-column view to save space on isolate view page

- give the dart:nativewrappers library a name

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org//262823012

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@35745 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/service.cc
@@ -1299,7 +1299,7 @@ static bool HandleScriptsEnumerate(Isolate* isolate, JSONStream* js) {
   JSONArray members(&jsobj, "members");
   const GrowableObjectArray& libs =
       GrowableObjectArray::Handle(isolate->object_store()->libraries());
-  int num_libs = libs.Length();
+  intptr_t num_libs = libs.Length();
   Library &lib = Library::Handle();
   Script& script = Script::Handle();
   for (intptr_t i = 0; i < num_libs; i++) {
